bottom navigation bar fix
