+++ lib/app.dart
@@ -1,8 +1,9 @@
 import 'package:flutter/material.dart';
-import 'package:mood_tracker/new_screen.dart';
 import 'package:mood_tracker/pin_screen/provider/pin_notifier.dart';
 import 'package:mood_tracker/provider/settings_notifier.dart';
 import 'package:provider/provider.dart';
+
+import 'custom_navigation_bar.dart';
 
 class App extends StatelessWidget {
   const App({Key? key}) : super(key: key);
@@ -18,9 +19,9 @@ class App extends StatelessWidget {
           create: (_) => SettingsNotifier(),
         ),
       ],
-      child: MaterialApp(
+      child: const MaterialApp(
         debugShowCheckedModeBanner: false,
-        home: NewScreen(),
+        home: CustomNavigationBar(),
       ),
     );
   }
